Improve Quarkus distribution integration tests
