+++ .github/workflows/ci.yml
@@ -5,6 +5,11 @@ on: [push, pull_request]
 env:
   DEFAULT_JDK_VERSION: 11
 
+concurrency:
+  # Only run once for latest commit per ref and cancel other (previous) runs.
+  group: ci-keycloak-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   build:
     name: Build
@@ -403,9 +408,13 @@ jobs:
           java-version: ${{ env.DEFAULT_JDK_VERSION }}
       - name: Update maven settings
         run: mkdir -p ~/.m2 ; cp .github/settings.xml ~/.m2/
-      - name: Run Quarkus Tests
+
+      - name: Prepare the local distribution archives
+        run: mvn clean install -DskipTests -Pdistribution
+
+      - name: Run Quarkus Integration Tests
         run: |
-          mvn clean install -nsu -B -f quarkus/tests/pom.xml  | misc/log/trimmer.sh
+          mvn clean install -nsu -B -f quarkus/tests/pom.xml -Dkc.quarkus.tests.enable=true | misc/log/trimmer.sh
           TEST_RESULT=${PIPESTATUS[0]}
           find . -path '*/target/surefire-reports/*.xml' | zip -q reports-quarkus-tests.zip -@
           exit $TEST_RESULT
